+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Changes in version 5.4**
 
 - Added NAACCR 24 documentation.
+- Switched CSV parser to the RFC4180-compliant one.
 - Updated NAACCR XML library from version 10.0 to version 10.1.
 - Updated CSV library from version 5.7.1 to version 5.8.